feat: add swagger docs

+++ src/exam-types/service/exams-type.service.ts
@@ -1,12 +1,16 @@
 import { Injectable } from '@nestjs/common';
 import { CreateExamTypeDto } from '../domain/dto/create-exam-type.dto';
 import { ExamTypes } from '@prisma/client';
-import examsTypeRepository from '../infra/exams-type.repository';
+import { ExamsTypeRepository } from '../infra/exams-type.repository';
 import { PaginationResponse } from 'src/core/utils/paginationResponse';
+import { AppoimentsService } from 'src/appoiments/services/appoiments.service';
 
 @Injectable()
 export class ExamsTypeService {
-  constructor(private examsTypeRepository: examsTypeRepository) {}
+  constructor(
+    private examsTypeRepository: ExamsTypeRepository,
+    private readonly appoimentService: AppoimentsService,
+  ) {}
 
   async createExamType(examType: CreateExamTypeDto): Promise<ExamTypes> {
     return this.examsTypeRepository.createExamType(examType);
@@ -17,6 +21,14 @@ export class ExamsTypeService {
   }
 
   async remove(id: string) {
+    const appoimentsWithExamType =
+      await this.appoimentService.findAllAppoimentsByExamTypeId(id);
+
+    if (appoimentsWithExamType > 0) {
+      throw new Error(
+        'Não é possível excluir este tipo de exame, pois existem agendamentos associados a ele.',
+      );
+    }
     return await this.examsTypeRepository.remove(id);
   }
